Refact Separação classes construtoras
