+++ constructors/__init__.py
@@ -1,0 +1,7 @@
+from constructors.classes import Cano, Passaro, Chao
+
+__all__ = [
+    "Cano", 
+    "Chao",
+    "Passaro"
+]
